Refactor XML parsing to use tinyxml2 library
Replaced wxXml library with tinyxml2 for XML parsing in pvxml_reader.cpp and pvxml_reader.h. Updated PVXMLReader methods and constructor to use tinyxml2 types and methods. Refactored code for readability and consistency, including using std::string instead of wxString, adding spaces around operators, and aligning code. Updated various methods to use static_cast for type safety. Renamed KernelExecutor::GetMessage to GetInfo and updated OclDlg::Execute accordingly. Improved type handling in VolumeBaker, VolumeSampler, and flrd namespace methods.

+++ FluoRender/FluoRender/Calculate/VolumeBaker.cpp
@@ -74,9 +74,9 @@ void VolumeBaker::Bake(bool replace)
 		return;
 
 	//input size
-	m_nx = input_nrrd->axis[0].size;
-	m_ny = input_nrrd->axis[1].size;
-	m_nz = input_nrrd->axis[2].size;
+	m_nx = static_cast<int>(input_nrrd->axis[0].size);
+	m_ny = static_cast<int>(input_nrrd->axis[1].size);
+	m_nz = static_cast<int>(input_nrrd->axis[2].size);
 	//bits
 	switch (input_nrrd->type)
 	{
